chore: remove sklearn from the path since the metrics should generalize to non-sklearn models
