+++ validmind/unit_metrics/classification/F1.py
@@ -2,12 +2,12 @@
 # See the LICENSE file in the root of this repository for details.
 # SPDX-License-Identifier: AGPL-3.0 AND ValidMind Commercial
 
-import numpy as np
+from sklearn.metrics import f1_score
 
 from validmind import tags, tasks
 
 
-@tags("regression")
-@tasks("regression")
-def MeanBiasDeviation(model, dataset):
-    return np.mean(dataset.y - dataset.y_pred(model))
+@tasks("classification")
+@tags("classification")
+def F1(model, dataset, **kwargs):
+    return f1_score(dataset.y, dataset.y_pred(model), **kwargs)
