+++ validmind/unit_metrics/classification/ROC_AUC.py
@@ -9,8 +9,8 @@ from sklearn.preprocessing import LabelBinarizer
 from validmind import tags, tasks
 
 
-@tags("classification", "sklearn", "unit_metric")
 @tasks("classification")
+@tags("classification")
 def ROC_AUC(model, dataset, **kwargs):
 
     y_true = dataset.y
