+++ validmind/unit_metrics/classification/Recall.py
@@ -2,13 +2,12 @@
 # See the LICENSE file in the root of this repository for details.
 # SPDX-License-Identifier: AGPL-3.0 AND ValidMind Commercial
 
-from sklearn.metrics import accuracy_score
+from sklearn.metrics import recall_score
 
 from validmind import tags, tasks
 
 
-@tags("classification", "sklearn", "unit_metric")
 @tasks("classification")
-def Accuracy(dataset, model):
-    """Calculates the accuracy of a model"""
-    return accuracy_score(dataset.y, dataset.y_pred(model))
+@tags("classification")
+def Recall(model, dataset, **kwargs):
+    return recall_score(dataset.y, dataset.y_pred(model), **kwargs)
