+++ validmind/unit_metrics/regression/HuberLoss.py
@@ -7,7 +7,7 @@ import numpy as np
 from validmind import tags, tasks
 
 
-@tags("regression", "unit_metric")
+@tags("regression")
 @tasks("regression")
 def HuberLoss(model, dataset):
     y_true = dataset.y
