+++ validmind/unit_metrics/regression/MeanAbsolutePercentageError.py
@@ -7,7 +7,7 @@ import numpy as np
 from validmind import tags, tasks
 
 
-@tags("regression", "unit_metric")
+@tags("regression")
 @tasks("regression")
 def MeanAbsolutePercentageError(model, dataset):
     y_true = dataset.y
