+++ validmind/unit_metrics/regression/MeanSquaredError.py
@@ -2,12 +2,12 @@
 # See the LICENSE file in the root of this repository for details.
 # SPDX-License-Identifier: AGPL-3.0 AND ValidMind Commercial
 
-from sklearn.metrics import mean_absolute_error as _mean_absolute_error
+from sklearn.metrics import mean_squared_error
 
 from validmind import tags, tasks
 
 
-@tags("regression", "sklearn", "unit_metric")
+@tags("regression")
 @tasks("regression")
-def MeanAbsoluteError(model, dataset, **kwargs):
-    return _mean_absolute_error(dataset.y, dataset.y_pred(model), **kwargs)
+def MeanSquaredError(model, dataset, **kwargs):
+    return mean_squared_error(dataset.y, dataset.y_pred(model), **kwargs)
