+++ validmind/unit_metrics/regression/RSquaredScore.py
@@ -2,12 +2,12 @@
 # See the LICENSE file in the root of this repository for details.
 # SPDX-License-Identifier: AGPL-3.0 AND ValidMind Commercial
 
-from sklearn.metrics import recall_score
+from sklearn.metrics import r2_score
 
 from validmind import tags, tasks
 
 
-@tags("classification", "sklearn", "unit_metric")
-@tasks("classification")
-def Recall(model, dataset, **kwargs):
-    return recall_score(dataset.y, dataset.y_pred(model), **kwargs)
+@tags("regression")
+@tasks("regression")
+def RSquaredError(model, dataset):
+    return r2_score(dataset.y, dataset.y_pred(model))
