+++ validmind/unit_metrics/regression/RootMeanSquaredError.py
@@ -8,7 +8,7 @@ from sklearn.metrics import mean_squared_error
 from validmind import tags, tasks
 
 
-@tags("regression", "sklearn", "unit_metric")
+@tags("regression")
 @tasks("regression")
 def RootMeanSquaredError(model, dataset, **kwargs):
     return np.sqrt(
